Move a yaw optimization function to Yaw module
